drawing, audio and UART work
